requirements.txt
anthropic>=0.40.0
openai>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
